Move StableGraph to its own file
